{"types":"C","runs":6}

+++ ConsoleProgrammSample.java
@@ -5,7 +5,7 @@ import acm.program.*;
 public class ConsoleProgrammSample extends ConsoleProgram {
 	public void run() {
 		println("This program creates the table of celsious to fahrenheit convention ");
-		for(int c = FROM; c<= TO; c += STEP) {
+		for(double c = FROM; c<= TO; c += STEP) {
 			println(c + "C = " + celsiousToFahrenheit(c) + "F");
 		}
 	}
